refactor: move types back into backend
